+++ kotlintest-runner-junit5/src/main/kotlin/io/kotlintest/runner/junit5/TestCaseRunner.kt
@@ -19,9 +19,14 @@ fun <CONTEXT> createInterceptorChain(
 
 class TestCaseRunner(private val listener: EngineExecutionListener) {
 
+  init {
+    println("created $this")
+  }
+
   // TODO beautify
   fun runTest(descriptor: TestCaseDescriptor) {
     if (descriptor.testCase.isActive()) {
+      println("Test is active")
       val executor =
           if (descriptor.testCase.config.threads < 2) Executors.newSingleThreadExecutor()
           else Executors.newFixedThreadPool(descriptor.testCase.config.threads)