+++ kotlintest-runner-junit5/src/test/java/qwe/qweqe/WibbleTest.java
@@ -1,0 +1,4 @@
+package qwe.qweqe;
+
+public class WibbleTest {
+}