+++ kotlintest-runner-junit5/src/test/kotlin/io/kotlintest/SharedInstanceForTestsTest.kt
@@ -12,11 +12,11 @@ class SharedInstanceForTestsTest : ShouldSpec() {
       count shouldBe 0
       count = 100
     }
-    should("be 0") {
+    should("be 100") {
       count shouldBe 100
       count = 200
     }
-    should("be 100") {
+    should("be 200") {
       count shouldBe 200
     }
   }