+++ kotlintest-runner-junit5/src/test/kotlin/io/kotlintest/matchers/MapMatchersTest.kt
@@ -37,5 +37,43 @@ class MapMatchersTest : WordSpec() {
         }
       }
     }
+
+    "containAll" should {
+      "test that a map contains all given pairs" {
+        val map = mapOf(Pair(1, "a"), Pair(2, "b"), Pair(3, "c"))
+        map should containAll(mapOf(1 to "a", 3 to "c"))
+        shouldThrow<AssertionError> {
+          map should containAll(mapOf(3 to "c"))
+        }
+      }
+    }
+
+    "haveKeys" should {
+      "test that a map contains all given keys" {
+        val map = mapOf(Pair(1, "a"), Pair(2, "b"), Pair(3, "c"))
+        map should haveKeys(1, 3)
+        map should haveKeys(1, 2, 3)
+        shouldThrow<AssertionError> {
+          map should haveKeys(4)
+        }
+        shouldThrow<AssertionError> {
+          map should haveKeys(1, 4)
+        }
+      }
+    }
+
+    "haveValues" should {
+      "test that a map contains all given values" {
+        val map = mapOf(Pair(1, "a"), Pair(2, "b"), Pair(3, "c"))
+        map should haveValues("b", "c")
+        map should haveValues("a", "b", "c")
+        shouldThrow<AssertionError> {
+          map should haveValues("a", "d")
+        }
+        shouldThrow<AssertionError> {
+          map should haveValues("d")
+        }
+      }
+    }
   }
 }