+++ kotlintest-runner-junit5/src/test/kotlin/io/kotlintest/specs/BehaviourSpecLambdaTest.kt
@@ -1,4 +1,4 @@
-package io.kotlintest.core
+package io.kotlintest.specs
 
 import io.kotlintest.shouldBe
 import io.kotlintest.specs.BehaviorSpec